requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
torch==1.9.0
transformers==4.11.3
googletrans==3.1.0a0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
aiohttp==3.8.1
requests==2.26.0
python-dotenv==0.19.0 